+++ src/inputs/types.d.ts
@@ -1,7 +1,7 @@
 import type { SupportedSvgoVersions } from "../svgo";
 
 interface Config {
-  readonly ignoreGlob: string;
+  readonly ignoreGlobs: string[];
   readonly isDryRun: boolean;
   readonly svgoConfigPath: string;
   readonly svgoVersion: SupportedSvgoVersions;