+++ test/unit/inputs/getters.test.ts
@@ -5,20 +5,18 @@ import inp from "../../__common__/inputter.mock";
 jest.mock("../../../src/errors");
 
 import {
-  getIgnoreGlob,
+  getIgnoreGlobs,
   getIsDryRun,
   getSvgoConfigPath,
   getSvgoVersion,
 } from "../../../src/inputs/getters";
 
 describe("inputs/getters.ts", () => {
-  const expectedGetInputOptions = { required: true };
-
   beforeEach(() => {
     resetAllWhenMocks();
   });
 
-  describe("::getIgnoreGlob", () => {
+  describe("::getIgnoreGlobs", () => {
     const inputKey = "ignore";
 
     beforeEach(() => {
@@ -28,14 +26,33 @@ describe("inputs/getters.ts", () => {
     test.each([
       "foobar",
       "Hello world!",
-    ])("can get input ('%s')", (configuredValue) => {
+    ])("can get input, single line ('%s')", (configuredValue) => {
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockReturnValueOnce(configuredValue);
+        .mockReturnValue(configuredValue);
 
-      const [result, err] = getIgnoreGlob(inp, "foobar");
+      const [result, err] = getIgnoreGlobs(inp, "foobar");
       expect(err).toBeNull();
-      expect(result).toEqual(configuredValue);
+      expect(result).toEqual([configuredValue]);
+    });
+
+    test.each([
+      [
+        "foo\nbar",
+        ["foo", "bar"],
+      ],
+      [
+        "Hello\r\nworld!",
+        ["Hello", "world!"],
+      ],
+    ])("can get input, multi line (%#)", (configuredValues, expectedValues) => {
+      when(inp.getInput)
+        .calledWith(inputKey, expect.anything())
+        .mockReturnValue(configuredValues);
+
+      const [result, err] = getIgnoreGlobs(inp, "foobar");
+      expect(err).toBeNull();
+      expect(result).toEqual(expectedValues);
     });
 
     test("can't get input", () => {
@@ -43,24 +60,11 @@ describe("inputs/getters.ts", () => {
 
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockImplementationOnce(() => { throw new Error(""); });
+        .mockImplementation(() => { throw new Error(""); });
 
-      const [result, err] = getIgnoreGlob(inp, defaultValue);
+      const [result, err] = getIgnoreGlobs(inp, defaultValue);
       expect(err).toBeNull();
-      expect(result).toEqual(defaultValue);
-    });
-
-    test("gets the input once", () => {
-      getIgnoreGlob(inp, "foobar");
-      expect(inp.getInput).toHaveBeenCalledTimes(1);
-    });
-
-    test("gets the input as being required", () => {
-      getIgnoreGlob(inp, "foobar");
-      expect(inp.getInput).toHaveBeenCalledWith(
-        expect.any(String),
-        expectedGetInputOptions,
-      );
+      expect(result).toEqual([defaultValue]);
     });
   });
 
@@ -77,7 +81,7 @@ describe("inputs/getters.ts", () => {
     ])("can get input (`%p`)", (configuredValue) => {
       when(inp.getBooleanInput)
         .calledWith(inputKey, expect.anything())
-        .mockReturnValueOnce(configuredValue);
+        .mockReturnValue(configuredValue);
 
       const [result, err] = getIsDryRun(inp, !configuredValue);
       expect(err).toBeNull();
@@ -89,24 +93,12 @@ describe("inputs/getters.ts", () => {
 
       when(inp.getBooleanInput)
         .calledWith(inputKey, expect.anything())
-        .mockImplementationOnce(() => { throw new Error(""); });
+        .mockImplementation(() => { throw new Error(""); });
 
       const [result, err] = getIsDryRun(inp, defaultValue);
-      expect(err).toBeNull();
-      expect(result).toEqual(defaultValue);
-    });
-
-    test("gets the input once", () => {
-      getIsDryRun(inp, false);
-      expect(inp.getBooleanInput).toHaveBeenCalledTimes(1);
-    });
-
-    test("gets the input as being required", () => {
-      getIsDryRun(inp, false);
-      expect(inp.getBooleanInput).toHaveBeenCalledWith(
-        expect.any(String),
-        expectedGetInputOptions,
-      );
+      expect(err).not.toBeNull();
+      expect(err).toContain("invalid dry-run value");
+      expect(result).toEqual(true);
     });
   });
 
@@ -123,7 +115,7 @@ describe("inputs/getters.ts", () => {
     ])("can get input ('%s')", (configuredValue) => {
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockReturnValueOnce(configuredValue);
+        .mockReturnValue(configuredValue);
 
       const [result, err] = getSvgoConfigPath(inp, "foo.bar");
       expect(err).toBeNull();
@@ -135,24 +127,11 @@ describe("inputs/getters.ts", () => {
 
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockImplementationOnce(() => { throw new Error(""); });
+        .mockImplementation(() => { throw new Error(""); });
 
       const [result, err] = getSvgoConfigPath(inp, defaultValue);
       expect(err).toBeNull();
       expect(result).toEqual(defaultValue);
-    });
-
-    test("gets the input once", () => {
-      getSvgoConfigPath(inp, "foo.bar");
-      expect(inp.getInput).toHaveBeenCalledTimes(1);
-    });
-
-    test("gets the input as being required", () => {
-      getSvgoConfigPath(inp, "foo.bar");
-      expect(inp.getInput).toHaveBeenCalledWith(
-        expect.any(String),
-        expectedGetInputOptions,
-      );
     });
   });
 
@@ -169,7 +148,7 @@ describe("inputs/getters.ts", () => {
     ])("can get input, valid (`%d`)", (configuredValue) => {
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockReturnValueOnce(`${configuredValue}`);
+        .mockReturnValue(`${configuredValue}`);
 
       const [result, err] = getSvgoVersion(inp, configuredValue === 1 ? 2 : 1);
       expect(err).toBeNull();
@@ -177,14 +156,31 @@ describe("inputs/getters.ts", () => {
     });
 
     test.each([
+      "0",
+      "3",
       "42",
+    ])("can get input, unsupported ('%s')", (configuredValue) => {
+      const defaultValue = 2;
+
+      when(inp.getInput)
+        .calledWith(inputKey, expect.anything())
+        .mockReturnValue(configuredValue);
+
+      const [result, err] = getSvgoVersion(inp, defaultValue);
+      expect(err).not.toBeNull();
+      expect(err).toContain("unsupported SVGO version");
+      expect(result).toEqual(defaultValue);
+    });
+
+    test.each([
       "foobar",
+      "Hello world",
     ])("can get input, invalid ('%s')", (configuredValue) => {
       const defaultValue = 2;
 
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockReturnValueOnce(configuredValue);
+        .mockReturnValue(configuredValue);
 
       const [result, err] = getSvgoVersion(inp, defaultValue);
       expect(err).not.toBeNull();
@@ -197,24 +193,11 @@ describe("inputs/getters.ts", () => {
 
       when(inp.getInput)
         .calledWith(inputKey, expect.anything())
-        .mockImplementationOnce(() => { throw new Error(""); });
+        .mockImplementation(() => { throw new Error(""); });
 
       const [result, err] = getSvgoVersion(inp, defaultValue);
       expect(err).toBeNull();
       expect(result).toEqual(defaultValue);
-    });
-
-    test("gets the input once", () => {
-      getSvgoVersion(inp, 2);
-      expect(inp.getInput).toHaveBeenCalledTimes(1);
-    });
-
-    test("gets the input as being required", () => {
-      getSvgoVersion(inp, 2);
-      expect(inp.getInput).toHaveBeenCalledWith(
-        expect.any(String),
-        expectedGetInputOptions,
-      );
     });
   });
 });